+++ example/logic_test.sh
@@ -12,7 +12,7 @@ function test_text_should_contain() {
 }
 
 function test_text_should_not_contain() {
-  assertNotContains "expecs" "$($SCRIPT "123")"
+  assert_not_contains "expecs" "$($SCRIPT "123")"
 }
 
 function test_text_should_match_a_regular_expression() {